[NUI.Components] Improve sam score to 4.21
