+++ src/Tizen.NUI.Components/Utils/StyleManager.cs
@@ -240,11 +240,13 @@ namespace Tizen.NUI.Components
 
         private void SetInitialThemeByDeviceProfile()
         {
-            themeMap.Add(wearableThemeName, WearableTheme.Instance);
-            themeMap.Add(defaultThemeName, DefaultTheme.Instance);
+            Theme wearableTheme = WearableTheme.Instance;
+            Theme defaultTheme = DefaultTheme.Instance;
+            themeMap.Add(wearableThemeName, wearableTheme);
+            themeMap.Add(defaultThemeName, defaultTheme);
 
             currentThemeName = defaultThemeName;
-            currentTheme = DefaultTheme.Instance;
+            currentTheme = defaultTheme;
 
             string currentProfile;
 
@@ -262,7 +264,7 @@ namespace Tizen.NUI.Components
             if (string.Equals(currentProfile, wearableThemeName))
             {
                 currentThemeName = wearableThemeName;
-                currentTheme = WearableTheme.Instance;
+                currentTheme = wearableTheme;
             }
         }
